pre-commit: Add `--overwrite` and `--local` to default args
Without `--overwrite`, the pre-commit hook does not work at all. Output
is printed to stdout but never seen by the user. Note that this still
spams useless backup files in the git repository (the git staging area
already acts as backup).

The `--local` flag is useful because you usually want every contributor
to use the same settings, so the local flags allows putting a
`.latexindent.yaml` in the root of the git repository.